divided scraping operation to other file.
